allow editing completed roadmap items
